+++ block-os/src/components/editor/Editor.tsx
@@ -1,10 +1,10 @@
 import { useEffect, useState, useRef } from 'react'
 import { useEditor, EditorContent, Editor as TiptapEditor } from '@tiptap/react'
 import StarterKit from '@tiptap/starter-kit'
-import { BlockLink, BlockReference, searchBlocks } from '../editor/extensions'
+import { BlockLink, BlockReference, searchBlocks } from '../../editor/extensions'
 import { SuggestionMenu } from './SuggestionMenu'
-import { documentStore } from '../storage/documentStore'
-import { blockStore } from '../storage/blockStore'
+import { documentStore } from '../../storage/documentStore'
+import { blockStore } from '../../storage/blockStore'
 import './Editor.css'
 
 interface EditorProps {