+++ block-os/src/components/layout/Sidebar.tsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react'
-import { Project, projectStore } from '../storage/projectStore'
+import { Project, projectStore } from '../../storage/projectStore'
 import './Sidebar.css'
 
 interface SidebarProps {