+++ block-os/src/components/panel/BlockSpacePanel.tsx
@@ -1,6 +1,6 @@
 import { useState, useEffect } from 'react'
-import { Block, blockStore } from '../storage/blockStore'
-import { formatRelativeTime } from '../utils/date'
+import { Block, blockStore } from '../../storage/blockStore'
+import { formatRelativeTime } from '../../utils/date'
 import './BlockSpacePanel.css'
 
 export function BlockSpacePanel() {